+++ CHANGELOG.md
@@ -11,6 +11,7 @@ These are changes that will probably be included in the next release.
 ### Added
  * Add Null logger
  * Add NewNamed convenience function
+ * Add Context support
  
 ### Fixed